l2-testing/requirements.txt
azure-devops
python-dotenv
azure-mgmt-devtestlabs
